Sync checking/unchecking items with the backend
The items reducer now handles a set of PATCH_ITEM actions. This includes
setting a loading flag on each item (in the event that we want to show an
inline spinner). As a result, the reducer for PATCH_ITEM_SUCCESS and
PATCH_ITEM_FAILURE both need to remove the `loading` property of items.

This PATCH_ITEM set of actions is used by the (now asynchronous)
toggleChecked() action creator.