Switch back from edit mode after submitting edit
Unfortunately, this required switching away from using `button`s for Edit and
Cancel and using `a` tags. If they were left as buttons, react would get
confused for some reason and treat the edit button as a form submit button.
It's probably a race condition, so swapping to an anchor link is the correct
way.

Anchor links were styled to look like buttons. The use of the .button class
might have side effects elsewhere in the code base however.