Fix delete button regression
My bad, it was a mistake to nest a button within a link anyway, and not valid
HTML5 markup. For the behavior of a clickable card, we have a number of future
options [as discussed in this post](https://inclusive-components.design/cards/).
None of which are perfect, but we should have a discussion about what the
correct behavior is (perhaps with wences).